# Frontend/backend_openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Valio Aimo Customer Portal API",
      "description": "Backend API for Zero-Fail Logistics Customer Portal",
      "version": "2.0.0"
    },
    "paths": {
...
      "/customer/dashboard/": {
        "get": {
          "tags": [
            "customer-dashboard"
          ],
          "summary": "Get Dashboard",
          "description": "Get customer dashboard with stats and recent activity",
          "operationId": "get_dashboard_customer_dashboard__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CustomerDashboard"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
      "/customer/dashboard/actions-needed": {
        "get": {
          "tags": [
            "customer-dashboard"
          ],
          "summary": "Get Actions Needed",
          "description": "Get orders that need customer action",
          "operationId": "get_actions_needed_customer_dashboard_actions_needed_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/Order"
                    },
                    "type": "array",
                    "title": "Response Get Actions Needed Customer Dashboard Actions Needed Get"
                  }
                }
              }
            }
          },
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ]
        }
      },
...
      "/admin/customers/": {
        "post": {
          "tags": [
            "admin-customers"
          ],
          "summary": "Create Customer",
          "description": "Create a new customer",
          "operationId": "create_customer_admin_customers__post",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerCreate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "admin-customers"
          ],
          "summary": "Get Customers",
          "description": "Get all customers",
          "operationId": "get_customers_admin_customers__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "skip",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 0,
                "title": "Skip"
              }
            },
            {
              "name": "limit",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "default": 100,
                "title": "Limit"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Customer"
                    },
                    "title": "Response Get Customers Admin Customers  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/admin/customers/{customer_id}": {
        "get": {
          "tags": [
            "admin-customers"
          ],
          "summary": "Get Customer",
          "description": "Get specific customer",
          "operationId": "get_customer_admin_customers__customer_id__get",
          "security": [
            {
              "OAuth2PasswordBearer": []
            }
          ],
          "parameters": [
            {
              "name": "customer_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "title": "Customer Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Claim": {
          "properties": {
            "claim_id": {
              "type": "string",
              "title": "Claim Id"
            },
            "order_id": {
              "type": "string",
              "title": "Order Id"
            },
            "customer_id": {
              "type": "string",
              "title": "Customer Id"
            },
            "claim_type": {
              "$ref": "#/components/schemas/ClaimType"
            },
            "status": {
              "$ref": "#/components/schemas/ClaimStatus"
            },
            "channel": {
              "type": "string",
              "title": "Channel"
            },
            "model_confidence_score": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Confidence Score"
            },
            "resolution_type": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ResolutionType"
                },
                {
                  "type": "null"
                }
              ]
            },
            "credit_amount": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Credit Amount"
            },
            "re_delivery_date": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Re Delivery Date"
            },
            "handled_by": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Handled By"
            },
            "rejection_reason": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rejection Reason"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            },
            "claim_lines": {
              "items": {
                "$ref": "#/components/schemas/ClaimLine"
              },
              "type": "array",
              "title": "Claim Lines",
              "default": []
            },
            "claim_attachments": {
              "items": {
                "$ref": "#/components/schemas/ClaimAttachment"
              },
              "type": "array",
              "title": "Claim Attachments",
              "default": []
            }
          },
          "type": "object",
          "required": [
            "claim_id",
            "order_id",
            "customer_id",
            "claim_type",
            "status",
            "channel",
            "created_at",
            "updated_at"
          ],
          "title": "Claim"
        },
        "ClaimAttachment": {
          "properties": {
            "attachment_id": {
              "type": "integer",
              "title": "Attachment Id"
            },
            "claim_id": {
              "type": "string",
              "title": "Claim Id"
            },
            "file_path": {
              "type": "string",
              "title": "File Path"
            },
            "file_type": {
              "type": "string",
              "title": "File Type"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            }
          },
          "type": "object",
          "required": [
            "attachment_id",
            "claim_id",
            "file_path",
            "file_type",
            "created_at"
          ],
          "title": "ClaimAttachment"
        },
        "ClaimLine": {
          "properties": {
            "claim_line_id": {
              "type": "integer",
              "title": "Claim Line Id"
            },
            "claim_id": {
              "type": "string",
              "title": "Claim Id"
            },
            "line_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Line Id"
            },
            "product_code": {
              "type": "string",
              "title": "Product Code"
            },
            "reported_issue": {
              "type": "string",
              "title": "Reported Issue"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            }
          },
          "type": "object",
          "required": [
            "claim_line_id",
            "claim_id",
            "product_code",
            "reported_issue",
            "created_at"
          ],
          "title": "ClaimLine"
        },
...
        "ClaimStatus": {
          "type": "string",
          "enum": [
            "open",
            "ai_processing",
            "manual_review",
            "approved",
            "rejected",
            "resolved"
          ],
          "title": "ClaimStatus"
        },
        "ClaimType": {
          "type": "string",
          "enum": [
            "MISSING_ITEM",
            "DAMAGED_ITEM",
            "WRONG_ITEM",
            "QUALITY_ISSUE"
          ],
          "title": "ClaimType"
        },
...
        "ContactChannel": {
          "type": "string",
          "enum": [
            "voice_first",
            "sms_first",
            "email_only"
          ],
          "title": "ContactChannel"
        },
        "Customer": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "segment": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Segment"
            },
            "language_preference": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Language"
                }
              ],
              "default": "fi"
            },
            "contact_channel": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ContactChannel"
                }
              ],
              "default": "voice_first"
            },
            "no_call_before": {
              "type": "string",
              "title": "No Call Before",
              "default": "07:00"
            },
            "accept_auto_substitutions_for_flexible_items": {
              "type": "boolean",
              "title": "Accept Auto Substitutions For Flexible Items",
              "default": true
            },
            "location": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            },
            "customer_id": {
              "type": "string",
              "title": "Customer Id"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            }
          },
          "type": "object",
          "required": [
            "name",
            "customer_id",
            "created_at",
            "updated_at"
          ],
          "title": "Customer"
        },
        "CustomerCreate": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "segment": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Segment"
            },
            "language_preference": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Language"
                }
              ],
              "default": "fi"
            },
            "contact_channel": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ContactChannel"
                }
              ],
              "default": "voice_first"
            },
            "no_call_before": {
              "type": "string",
              "title": "No Call Before",
              "default": "07:00"
            },
            "accept_auto_substitutions_for_flexible_items": {
              "type": "boolean",
              "title": "Accept Auto Substitutions For Flexible Items",
              "default": true
            },
            "location": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            },
            "customer_id": {
              "type": "string",
              "title": "Customer Id"
            }
          },
          "type": "object",
          "required": [
            "name",
            "customer_id"
          ],
          "title": "CustomerCreate"
        },
        "CustomerDashboard": {
          "properties": {
            "customer": {
              "$ref": "#/components/schemas/Customer"
            },
            "recent_orders": {
              "items": {
                "$ref": "#/components/schemas/Order"
              },
              "type": "array",
              "title": "Recent Orders"
            },
            "open_claims": {
              "items": {
                "$ref": "#/components/schemas/Claim"
              },
              "type": "array",
              "title": "Open Claims"
            },
            "stats": {
              "$ref": "#/components/schemas/OrderStats"
            }
          },
          "type": "object",
          "required": [
            "customer",
            "recent_orders",
            "open_claims",
            "stats"
          ],
          "title": "CustomerDashboard"
        },
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "Language": {
          "type": "string",
          "enum": [
            "fi",
            "sv",
            "en"
          ],
          "title": "Language"
        },
        "LineStatus": {
          "type": "string",
          "enum": [
            "OK",
            "PARTIAL",
            "ZERO",
            "REPLACED"
          ],
          "title": "LineStatus"
        },
...
        "Order": {
          "properties": {
            "delivery_date": {
              "type": "string",
              "title": "Delivery Date"
            },
            "delivery_window_start": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delivery Window Start"
            },
            "delivery_window_end": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Delivery Window End"
            },
            "channel": {
              "type": "string",
              "title": "Channel",
              "default": "web"
            },
            "order_id": {
              "type": "string",
              "title": "Order Id"
            },
            "customer_id": {
              "type": "string",
              "title": "Customer Id"
            },
            "order_datetime": {
              "type": "string",
              "format": "date-time",
              "title": "Order Datetime"
            },
            "status": {
              "$ref": "#/components/schemas/OrderStatus"
            },
            "overall_order_risk": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Overall Order Risk"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            },
            "order_lines": {
              "items": {
                "$ref": "#/components/schemas/OrderLine"
              },
              "type": "array",
              "title": "Order Lines",
              "default": []
            }
          },
          "type": "object",
          "required": [
            "delivery_date",
            "order_id",
            "customer_id",
            "order_datetime",
            "status",
            "created_at",
            "updated_at"
          ],
          "title": "Order"
        },
        "OrderChange": {
          "properties": {
            "change_id": {
              "type": "integer",
              "title": "Change Id"
            },
            "order_id": {
              "type": "string",
              "title": "Order Id"
            },
            "line_id": {
              "type": "integer",
              "title": "Line Id"
            },
            "old_product_code": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Old Product Code"
            },
            "new_product_code": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "New Product Code"
            },
            "change_reason": {
              "type": "string",
              "title": "Change Reason"
            },
            "confirmed_by": {
              "type": "string",
              "title": "Confirmed By"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            }
          },
          "type": "object",
          "required": [
            "change_id",
            "order_id",
            "line_id",
            "change_reason",
            "confirmed_by",
            "created_at"
          ],
          "title": "OrderChange"
        },
        "OrderLine": {
          "properties": {
            "product_code": {
              "type": "string",
              "title": "Product Code"
            },
            "ordered_qty": {
              "type": "number",
              "title": "Ordered Qty"
            },
            "item_priority": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/PriorityLevel"
                }
              ],
              "default": "IMPORTANT"
            },
            "auto_substitution_allowed": {
              "type": "boolean",
              "title": "Auto Substitution Allowed",
              "default": true
            },
            "customer_comments": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Comments"
            },
            "line_id": {
              "type": "integer",
              "title": "Line Id"
            },
            "order_id": {
              "type": "string",
              "title": "Order Id"
            },
            "delivered_qty": {
              "type": "number",
              "title": "Delivered Qty"
            },
            "shortage_flag": {
              "type": "boolean",
              "title": "Shortage Flag"
            },
            "shortage_ratio": {
              "type": "number",
              "title": "Shortage Ratio"
            },
            "risk_score": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Score"
            },
            "line_status": {
              "$ref": "#/components/schemas/LineStatus"
            },
            "product": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Product"
                },
                {
                  "type": "null"
                }
              ]
            },
            "order_changes": {
              "items": {
                "$ref": "#/components/schemas/OrderChange"
              },
              "type": "array",
              "title": "Order Changes",
              "default": []
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            }
          },
          "type": "object",
          "required": [
            "product_code",
            "ordered_qty",
            "line_id",
            "order_id",
            "delivered_qty",
            "shortage_flag",
            "shortage_ratio",
            "line_status",
            "created_at",
            "updated_at"
          ],
          "title": "OrderLine"
        },
...
        "OrderStats": {
          "properties": {
            "total_orders": {
              "type": "integer",
              "title": "Total Orders"
            },
            "orders_at_risk": {
              "type": "integer",
              "title": "Orders At Risk"
            },
            "orders_in_picking": {
              "type": "integer",
              "title": "Orders In Picking"
            },
            "orders_delivered_today": {
              "type": "integer",
              "title": "Orders Delivered Today"
            }
          },
          "type": "object",
          "required": [
            "total_orders",
            "orders_at_risk",
            "orders_in_picking",
            "orders_delivered_today"
          ],
          "title": "OrderStats"
        },
        "OrderStatus": {
          "type": "string",
          "enum": [
            "placed",
            "under_risk",
            "waiting_for_customer_action",
            "picking",
            "delivering",
            "delivered",
            "cancelled"
          ],
          "title": "OrderStatus"
        },
...
        "PriorityLevel": {
          "type": "string",
          "enum": [
            "CRITICAL",
            "IMPORTANT",
            "FLEXIBLE"
          ],
          "title": "PriorityLevel"
        },
        "Product": {
          "properties": {
            "product_name": {
              "type": "string",
              "title": "Product Name"
            },
            "category": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            },
            "sub_category": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sub Category"
            },
            "temperature_zone": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Temperature Zone"
            },
            "shelf_life_days": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shelf Life Days"
            },
            "unit_size": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unit Size"
            },
            "unit_type": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unit Type"
            },
            "price": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Price"
            },
            "product_code": {
              "type": "string",
              "title": "Product Code"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            }
          },
          "type": "object",
          "required": [
            "product_name",
            "product_code",
            "created_at",
            "updated_at"
          ],
          "title": "Product"
        },
...
        "ResolutionType": {
          "type": "string",
          "enum": [
            "CREDIT",
            "REDELIVERY_SAME_DAY",
            "REDELIVERY_NEXT_DAY",
            "REPLACEMENT_NEXT_ORDER"
          ],
          "title": "ResolutionType"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      },
      "securitySchemes": {
        "OAuth2PasswordBearer": {
          "type": "oauth2",
          "flows": {
            "password": {
              "scopes": {
  
              },
              "tokenUrl": "auth/login"
            }
          }
        }
      }
    }
  }